+++ test/helpers/api/public/content/experience/e-4b843ea93ed9d4.js
@@ -2,6 +2,7 @@ module.exports = {
     id: 'e-4b843ea93ed9d4',
     data: {
         title: 'MiniReel 2',
+        mode: 'embed',
         collateral: {
             splash: '/collateral/splash2.jpg'
         },