+++ test/helpers/api/public/content/experience/e-6b5ead50d4a1ed.js
@@ -1,6 +1,7 @@
 module.exports = {
     id: 'e-6b5ead50d4a1ed',
     data: {
+        mode: 'lightbox-playlist',
         title: 'MiniReel 3',
         collateral: {
             splash: '/collateral/splash1.jpg'