+++ test/helpers/api/public/content/experience/e-fcb95ef54b22f5.js
@@ -1,6 +1,7 @@
 module.exports = {
     id: 'e-fcb95ef54b22f5',
     data: {
+        mode: 'embed',
         title: 'MiniReel 1',
         collateral: {
             splash: '/collateral/splash3.jpg'