+++ test/helpers/collateral/mr2/templates/test.js
@@ -12,5 +12,9 @@ module.exports = [
     '        <h1>{{title}}</h1>',
     '        <img width="5" height="5" tb-src="{{splash}}" />',
     '    </div>',
+    '    <div class="c6-mr2__mr-splash">',
+    '        <h1>{{title}}</h1>',
+    '        <img width="5" height="5" tb-src="{{splash}}" />',
+    '    </div>',
     '</div>'
 ].join('\n');