+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # C6Embed Changelog
 
+* [FEATURE]: Pass container up to content service when getting experiences
+* [FEATURE]: Move lightbox override to jsonp/mr2 handlers, stop sending context to service
+
 ## v2.14.0 (January 23, 2015)
 * [FEATURE]: Making adtech network and server address somewhat more configurable